deployed

+++ task2/priv/repo/migrations/20180224154349_create_timeblocks.exs
@@ -3,8 +3,8 @@ defmodule Task2.Repo.Migrations.CreateTimeblocks do
 
   def change do
     create table(:timeblocks) do
-      add :start_time, :integer
-      add :end_time, :integer
+      add :start_time, :bigint
+      add :end_time, :bigint
       add :task_id, references(:tasks, on_delete: :delete_all), null: false
 
       timestamps()
